third_party/blink/web_tests/platform/generic/http/tests/css/performance-info-with-cached-sheet-expected.txt

Test for crbug.com/337065: resources in cached stylesheet should have resource timing and CDT timeline.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS window.performance.getEntriesByName("http://127.0.0.1:8000/css/resources/abe.png") is not []
PASS window.performance.getEntriesByName("http://127.0.0.1:8000/resources/square20.png") is []
PASS window.performance.getEntriesByName("http://127.0.0.1:8000/misc/resources/acid3/font.ttf") is not []
Acid Ahem Test Font
